Log-likelihood function of the (normally mixed) multinomial probit model

Log-likelihood function of the (normally mixed) multinomial probit model

Usage

f_ll_mnp(theta, data, neg = FALSE, normal_cdf = mvtnorm::pmvnorm)

Arguments

theta

A numeric, the vector of model coefficients, see the details.

data

A data.frame, the output of sim_mnp.

neg

Set to TRUE to return the negative log-likelihood value.

normal_cdf

A function that evaluates the CDF of an n-variate normal distribution. It must take the arguments

lower, the vector of lower limits of length n,

upper, the vector of upper limits of length n,

mean, the mean vector of length n,

sigma, the n times n covariance matrix,

and return a single numeric.

Value

A numeric, the log-likelihood value at theta.

Details

The order of theta is supposed to be c(b,o,l), where

b denotes the vector of mean effects without the first entry,

o the lower-triangular elements of the lower-triangular Cholesky root of the effect covariance matrix Omega (if any),

and l the lower-triangular elements of the lower-triangular Cholesky root L of the differenced (with respect to the last alternative) error term covariance matrix Sigma.

References

https://en.wikipedia.org/wiki/Multinomial_probit

See also

Examples

data <- sim_mnp(N = 200, J = 3, P = 2, b = c(1,3))
theta <- attr(data, "true")[-1]
f_ll_mnp(theta = theta, data = data)
#> [1] -56.7067
# \donttest{
nlm(f_ll_mnp, p = theta, data = data, neg = TRUE)$estimate
#> [1] 2.944223 1.007717 0.265125 1.121199
# }