Simulate data from a (normally mixed) multinomial probit model

Simulate data from a (normally mixed) multinomial probit model

Usage

sim_mnp(
  N,
  T = 1,
  J,
  P,
  b = stats::rnorm(P),
  Omega = NULL,
  Sigma = diag(J),
  X = function(n, t) matrix(stats::rnorm(J * P), nrow = J, ncol = P)
)

Arguments

N

An integer, the number of observations.

T

An integer, the number of choice occasions. By default, T = 1.

J

An integer, the number of alternatives.

P

An integer, the number of choice covariates.

b

A numeric vector, the mean effects vector of length P. The first element must be 1 (for normalization).

Omega

A matrix, the covariance matrix of the normal mixing distribution of dimension P times P. Set to NULL (the default) for no mixing distribution.

Sigma

A matrix, the error term covariance matrix of dimension J times J.

X

A function that samples the covariates of decider n at choice occasion t. It must

have two arguments n and t,

return a numeric matrix of dimension J times P.

Value

A data.frame. The first column (N) is the identifier for the decider, the next column (T) the identifier for the choice occasion. Next comes the column y with the indices of the chosen alternatives. The last columns contain the column-wise entries of the covariate matrices.

The true model coefficients are added to the output via the attribute "true". They are already normalized and can be directly compared with the maximum likelihood estimate.

Additional attributes are "J" (the number of alternatives), "P" (the number of choice covariates), and "mix" (a boolean which is TRUE if Omega is not NULL).

See also

Examples

sim_mnp(N = 3, J = 2, P = 2, b = c(1,-1), Omega = diag(2), Sigma = diag(2))
#>   N T y          1          2          3         4
#> 1 1 1 1 -0.2641145 -1.6843813 -2.3442969 0.3219130
#> 2 2 1 2 -1.0435238 -1.2559566  0.5038673 0.6550934
#> 3 3 1 2 -0.1246941 -0.4768646 -0.5785750 0.3330870